+++ bin/formatjson.ps1
@@ -18,7 +18,7 @@ param(
     [String] $App = '*',
     [Parameter(Mandatory = $true)]
     [ValidateScript( {
-        if (!(Test-Path $_ -Type Container)) {
+        if (-not (Test-Path $_ -Type Container)) {
             throw "$_ is not a directory!"
         } else {
             $true
@@ -34,11 +34,23 @@ param(
 $Dir = Convert-Path $Dir
 
 Get-ChildItem $Dir -Filter "$App.json" -Recurse | ForEach-Object {
-    $file = $_.FullName
-    # beautify
-    $json = parse_json $file | ConvertToPrettyJson
+    # Path of file
+    $file = [string] $_.'FullName'
 
-    # convert to 4 spaces
-    $json = $json -replace "`t", '    '
-    [System.IO.File]::WriteAllLines($file, $json)
+    # Parse JSON
+    $json = [PSCustomObject](parse_json -path $file)
+
+    # Sort JSON root properties according to schema.json, and level one child properties alphabetically
+    $json = [PSCustomObject](Sort-ScoopManifestProperties -JsonAsObject $json)
+
+    # Beautify
+    $json = [string](ConvertToPrettyJson -data $json)
+
+    # Convert to 4 spaces
+    $json = [string]($json -replace "`t", '    ')
+
+    # Overwrite file content
+    if (-not [string]::IsNullOrWhiteSpace($json)) {
+        [System.IO.File]::WriteAllLines($file, $json)
+    }
 }